refactor(report): the props contract and the dead-link page are not the render
480 -> 352, and ReportView's baseline entry is REMOVED rather than tightened.
813 lines of wall becomes 48 lines of headroom under the 400 limit, which is
what #20 and #21-#26 needed.

`report/report-view-props.ts` — the ReportViewProps contract and the pure
`reportViewProps()` adapter. This is the report's INPUT boundary, not part of
its render: three callers (the standalone route, the agent view of that same
route, and the inline Hub slot) each hold a loader payload plus a few route
params and all three go through this one function. Having it importable
without React is what already lets `report-view.test.ts` exercise it directly;
that spec keeps importing it from ReportView, which re-exports both symbols.

`report/ReportUnavailable.tsx` — the three outcomes that render INSTEAD of a
report: not published, dead link, generic failure. They belong together
because choosing between them is the entire logic, and the IA-36 reasoning for
why 410 and 404 deliberately render the same page travels with the branch it
explains rather than sitting in the middle of a component that otherwise draws
a report.

ReportView's own header comment now says what it keeps and why: the
report-wide state (filter, lightbox, repair selection, failed-photo Set), the
media-tile renderer those close over, and the composition order. Everything a
reader looks up by name is a file next door.
