+++ scripts/file-size-baseline.json
@@ -7,7 +7,6 @@
   "app/routes/inspections.tsx": 879,
   "server/lib/collab/results-doc.ts": 874,
   "server/api/sms.ts": 844,
-  "app/components/portal/sections/ReportView.tsx": 813,
   "app/routes/settings-communication.tsx": 777,
   "server/api/admin/admin-settings.ts": 754,
   "server/services/inspection.service.ts": 741,
